requirements.txt
fastapi
uvicorn
openai
python-dotenv
pydantic
langchain
langchain-community
langchain-text-splitters
sentence-transformers
faiss-cpu
numpy
python-docx